+++ test/integration/test_brute_force.py
@@ -37,6 +37,9 @@ class TestBruteForce(TestCaseIntegration):
         args = create_base_args()
         args.input_gvf = gvf
         args.reference_dir = reference
+        args.max_adjacent_as_mnv = 0
+        args.w2f_reassignment = False
+        args.selenocysteine_termination = False
         stream = io.StringIO()
         with contextlib.redirect_stdout(stream):
             util.brute_force(args)